feat: upgrades sentry to version 8
