+++ package.json
@@ -55,7 +55,7 @@
     "@radix-ui/react-tabs": "^1.0.4",
     "@radix-ui/react-toast": "^1.1.5",
     "@radix-ui/react-tooltip": "^1.0.7",
-    "@sentry/nextjs": "^7.116.0",
+    "@sentry/nextjs": "^8.7.0",
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "@tanstack/react-table": "^8.11.2",
     "@thirdweb-dev/auth": "^4.1.30",
